Update README.md
I am doing an Intro to Docker presentation for Python Developers for Ireland Python June meetup.  An important part of this presentation is a demo of the Python repository on DockerHub.  I am suggesting changes to the repository readme to make it easier to understand for Python developers who are new to Docker.  I am trying to give more examples of image name syntax and steer users towards using the version specific versions of the image needs.